Cleanup error classes and CI
