+++ .github/workflows/test.yml
@@ -1,0 +1,14 @@
+name: "Test"
+on: [push, pull_request] 
+
+jobs:
+  check_pr:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v1
+
+    - name: "yarn install"
+      run: yarn install
+
+    - name: "yarn test"
+      run: yarn test
