+++ lib/Action.js
@@ -10,7 +10,7 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.Action = void 0;
-const ErrorMessage_1 = require("./ErrorMessage");
+const GithubError_1 = require("./GithubError");
 class Action {
     constructor(inputs, releases, uploader) {
         this.inputs = inputs;
@@ -55,8 +55,8 @@ class Action {
         });
     }
     static noPublishedRelease(error) {
-        const errorMessage = new ErrorMessage_1.ErrorMessage(error);
-        return errorMessage.status == 404;
+        const githubError = new GithubError_1.GithubError(error);
+        return githubError.status == 404;
     }
     updateDraftOrCreateRelease() {
         return __awaiter(this, void 0, void 0, function* () {
