+++ lib/GithubError.js
@@ -1,57 +1,40 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.GithubError = void 0;
+const GithubErrorDetail_1 = require("./GithubErrorDetail");
 class GithubError {
     constructor(error) {
         this.error = error;
+        this.githubErrors = this.generateGithubErrors();
     }
-    get code() {
-        return this.error.code;
-    }
-    toString() {
-        const code = this.error.code;
-        switch (code) {
-            case 'missing':
-                return this.missingResourceMessage();
-            case 'missing_field':
-                return this.missingFieldMessage();
-            case 'invalid':
-                return this.invalidFieldMessage();
-            case 'already_exists':
-                return this.resourceAlreadyExists();
-            default:
-                return this.customErrorMessage();
-        }
-    }
-    customErrorMessage() {
-        const message = this.error.message;
-        const documentation = this.error.documentation_url;
-        let documentationMessage;
-        if (documentation) {
-            documentationMessage = `\nPlease see ${documentation}.`;
+    generateGithubErrors() {
+        const errors = this.error.errors;
+        if (errors instanceof Array) {
+            return errors.map((err) => new GithubErrorDetail_1.GithubErrorDetail(err));
         }
         else {
-            documentationMessage = "";
+            return [];
         }
-        return `${message}${documentationMessage}`;
     }
-    invalidFieldMessage() {
-        const resource = this.error.resource;
-        const field = this.error.field;
-        return `The ${field} field on ${resource} is an invalid format.`;
+    get status() {
+        return this.error.status;
     }
-    missingResourceMessage() {
-        const resource = this.error.resource;
-        return `${resource} does not exist.`;
+    hasErrorWithCode(code) {
+        return this.githubErrors.some((err) => err.code == code);
     }
-    missingFieldMessage() {
-        const resource = this.error.resource;
-        const field = this.error.field;
-        return `The ${field} field on ${resource} is missing.`;
+    toString() {
+        const message = this.error.message;
+        const errors = this.githubErrors;
+        const status = this.status;
+        if (errors.length > 0) {
+            return `Error ${status}: ${message}\nErrors:\n${this.errorBulletedList(errors)}`;
+        }
+        else {
+            return `Error ${status}: ${message}`;
+        }
     }
-    resourceAlreadyExists() {
-        const resource = this.error.resource;
-        return `${resource} already exists.`;
+    errorBulletedList(errors) {
+        return errors.map((err) => `- ${err}`).join("\n");
     }
 }
 exports.GithubError = GithubError;
