+++ lib/Main.js
@@ -35,7 +35,7 @@ const Releases_1 = require("./Releases");
 const Action_1 = require("./Action");
 const ArtifactUploader_1 = require("./ArtifactUploader");
 const ArtifactGlobber_1 = require("./ArtifactGlobber");
-const ErrorMessage_1 = require("./ErrorMessage");
+const GithubError_1 = require("./GithubError");
 function run() {
     return __awaiter(this, void 0, void 0, function* () {
         try {
@@ -43,8 +43,8 @@ function run() {
             yield action.perform();
         }
         catch (error) {
-            const errorMessage = new ErrorMessage_1.ErrorMessage(error);
-            core.setFailed(errorMessage.toString());
+            const githubError = new GithubError_1.GithubError(error);
+            core.setFailed(githubError.toString());
         }
     });
 }
@@ -55,7 +55,7 @@ function createAction() {
     const globber = new ArtifactGlobber_1.FileArtifactGlobber();
     const inputs = new Inputs_1.CoreInputs(globber, context);
     const releases = new Releases_1.GithubReleases(inputs, git);
-    const uploader = new ArtifactUploader_1.GithubArtifactUploader(releases, inputs.replacesArtifacts);
+    const uploader = new ArtifactUploader_1.GithubArtifactUploader(releases, inputs.replacesArtifacts, inputs.artifactErrorsFailBuild);
     return new Action_1.Action(inputs, releases, uploader);
 }
 run();
